feat: break out into seperate TC action
