+++ protocol-configuration/proposalsConfig.ts
@@ -3,6 +3,7 @@ import { ProposalCategory, TemplatedProposalsConfigMap } from '@custom-types/typ
 import tip_117 from '@proposals/description/tip_117';
 import pod_exec_v2 from '@proposals/description/pod_exec_v2';
 import tip_115 from '@proposals/description/tip_115';
+import tip_118 from '@proposals/description/tip_118';
 
 const proposals: TemplatedProposalsConfigMap = {
   tip_117: {
@@ -27,6 +28,15 @@ const proposals: TemplatedProposalsConfigMap = {
     deploy: false, // deploy flag for whether to run deploy action during e2e tests or use mainnet state
     totalValue: 0, // amount of ETH to send to DAO execution
     proposal: pod_exec_v2, // full proposal file, imported from '@proposals/description/fip_xx.ts'
+    proposalId: '',
+    affectedContractSignoff: [],
+    deprecatedContractSignoff: [],
+    category: ProposalCategory.TC
+  },
+  tip_118: {
+    deploy: false, // deploy flag for whether to run deploy action during e2e tests or use mainnet state
+    totalValue: 0, // amount of ETH to send to DAO execution
+    proposal: tip_118, // full proposal file, imported from '@proposals/description/fip_xx.ts'
     proposalId: '',
     affectedContractSignoff: [],
     deprecatedContractSignoff: [],
